refactor(monitor): 拆分自指标与模型探针子模块
